requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
pandas
numpy
requests
python-dotenv
